// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/project-search-result.model.ts
import { Project } from './project.model';

export class ProjectSearchResult extends Project {
  projectTypeName: string;
  projectStatusColor: string;
  userName: string;
  profileImage: string;
  createdDateTime: Date;
  fullName: string;
  roleName: string;
  projectResponsiblePersonId: string;
  activeGoalCount: number;
  numberOfReds: number;
  testSuiteCount: number;
  testRunCount: number;
  milestoneCount: number;
  totalCount: number;
  reportCount: number;
  casesCount: number;
  viewGoalsPermissionCount: number;
}